+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * 【setting】     修复utf8-bom的setting文件读取问题（issue#I7G34E@Gitee）
 * 【core  】      修复PathUtil.getMimeType可能造成的异常（issue#3179@Github）
 * 【core  】      修复Pair序列化转换无效问题（issue#I7GPGX@Github）
+* 【core  】      修复TypeUtil.getTypeArgument对实现接口获取不全面问题（issue#I7CRIW@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.20(2023-06-16)